Refactor text/chat completions APIs into _completion.py sub-module
