+++ client/h2ogpt_client/__init__.py
@@ -1,3 +1,4 @@
-from h2ogpt_client._core import Client, LangChainMode, PromptType
+from h2ogpt_client._core import Client
+from h2ogpt_client._h2ogpt_enums import LangChainMode, PromptType
 
 __all__ = ["Client", "PromptType", "LangChainMode"]
